Fix missing internal rids (#261)
- Also removed the auto updater thing.
- Rewrite how id's work internally, so now instead of handle we use a
global ID for all objects (handles werent globally unique).

+++ src/shapes/rapier_shape_base.rs
@@ -1,12 +1,24 @@
 use godot::prelude::*;
 use hashbrown::HashMap;
-use rapier::prelude::RigidBodyHandle;
+use rapier::prelude::SharedShape;
 
 use crate::bodies::rapier_collision_object::IRapierCollisionObject;
 use crate::rapier_wrapper::prelude::*;
-use crate::servers::rapier_physics_singleton::get_body_rid;
+use crate::servers::rapier_physics_singleton::get_id_rid;
+use crate::servers::rapier_physics_singleton::next_id;
 use crate::servers::rapier_physics_singleton::PhysicsData;
+use crate::servers::rapier_physics_singleton::RapierId;
 use crate::types::*;
+#[cfg_attr(feature = "serde-serialize", derive(serde::Serialize))]
+pub struct ShapeExport<'a> {
+    state: &'a RapierShapeState,
+    shape: &'a SharedShape,
+}
+#[cfg_attr(feature = "serde-serialize", derive(serde::Deserialize))]
+pub struct ShapeImport {
+    state: RapierShapeState,
+    shape: SharedShape,
+}
 #[cfg_attr(
     feature = "serde-serialize",
     derive(serde::Serialize, serde::Deserialize)
@@ -14,8 +26,16 @@ use crate::types::*;
 #[derive(PartialEq, Clone, Debug, Default)]
 pub struct RapierShapeState {
     aabb: Rect,
-    owners: HashMap<RigidBodyHandle, i32>,
+    #[cfg_attr(
+        feature = "serde-serialize",
+        serde(
+            serialize_with = "rapier::utils::serde::serialize_to_vec_tuple",
+            deserialize_with = "rapier::utils::serde::deserialize_from_vec_tuple"
+        )
+    )]
+    owners: HashMap<RapierId, i32>,
     handle: ShapeHandle,
+    id: RapierId,
 }
 pub struct RapierShapeBase {
     rid: Rid,
@@ -33,7 +53,10 @@ impl RapierShapeBase {
     pub(super) fn new(rid: Rid) -> Self {
         Self {
             rid,
-            state: RapierShapeState::default(),
+            state: RapierShapeState {
+                id: next_id(),
+                ..Default::default()
+            },
         }
     }
 
@@ -42,6 +65,11 @@ impl RapierShapeBase {
         handle: ShapeHandle,
         physics_engine: &mut PhysicsEngine,
     ) {
+        // new handle has to be valid
+        if handle == ShapeHandle::default() {
+            godot_error!("Invalid shape handle");
+            return;
+        }
         // destroy previous shape
         if self.state.handle != ShapeHandle::default() {
             self.destroy_shape(physics_engine);
@@ -64,21 +92,22 @@ impl RapierShapeBase {
     }
 
     pub fn call_shape_changed(
-        owners: HashMap<RigidBodyHandle, i32>,
-        shape_rid: Rid,
+        owners: HashMap<RapierId, i32>,
+        shape_id: RapierId,
+        shape_handle: ShapeHandle,
         physics_data: &mut PhysicsData,
     ) {
         for (owner, _) in owners {
             if let Some(owner) = physics_data
                 .collision_objects
-                .get_mut(&get_body_rid(owner, &physics_data.rids))
+                .get_mut(&get_id_rid(owner, &physics_data.ids))
             {
                 owner.shape_changed(
-                    shape_rid,
+                    shape_id,
+                    shape_handle,
                     &mut physics_data.physics_engine,
-                    &mut physics_data.shapes,
                     &mut physics_data.spaces,
-                    &physics_data.rids,
+                    &physics_data.ids,
                 );
             }
         }
@@ -90,11 +119,11 @@ impl RapierShapeBase {
         aabb_clone
     }
 
-    pub fn add_owner(&mut self, owner: RigidBodyHandle) {
+    pub fn add_owner(&mut self, owner: RapierId) {
         *self.state.owners.entry(owner).or_insert(0) += 1;
     }
 
-    pub fn remove_owner(&mut self, owner: RigidBodyHandle) {
+    pub fn remove_owner(&mut self, owner: RapierId) {
         if let Some(count) = self.state.owners.get_mut(&owner) {
             *count -= 1;
             if *count == 0 {
@@ -103,12 +132,16 @@ impl RapierShapeBase {
         }
     }
 
-    pub fn get_owners(&self) -> &HashMap<RigidBodyHandle, i32> {
+    pub fn get_owners(&self) -> &HashMap<RapierId, i32> {
         &self.state.owners
     }
 
     pub fn get_rid(&self) -> Rid {
         self.rid
+    }
+
+    pub fn get_id(&self) -> RapierId {
+        self.state.id
     }
 
     pub fn destroy_shape(&mut self, physics_engine: &mut PhysicsEngine) {
@@ -130,27 +163,34 @@ impl RapierShapeBase {
     }
 
     #[cfg(feature = "serde-serialize")]
-    pub fn export_binary(&self) -> PackedByteArray {
+    pub fn export_binary(&self, physics_engine: &mut PhysicsEngine) -> PackedByteArray {
         let mut buf = PackedByteArray::new();
-        match bincode::serialize(&self.state) {
-            Ok(binary_data) => {
-                buf.resize(binary_data.len());
-                for i in 0..binary_data.len() {
-                    buf[i] = binary_data[i];
+        if let Some(inner) = physics_engine.get_shape(self.get_handle()) {
+            let export = ShapeExport {
+                state: &self.state,
+                shape: inner,
+            };
+            match bincode::serialize(&export) {
+                Ok(binary_data) => {
+                    buf.resize(binary_data.len());
+                    for i in 0..binary_data.len() {
+                        buf[i] = binary_data[i];
+                    }
                 }
-            }
-            Err(e) => {
-                godot_error!("Failed to serialize shape to binary: {}", e);
+                Err(e) => {
+                    godot_error!("Failed to serialize shape to binary: {}", e);
+                }
             }
         }
         buf
     }
 
     #[cfg(feature = "serde-serialize")]
-    pub fn import_binary(&mut self, data: PackedByteArray) {
-        match bincode::deserialize::<RapierShapeState>(data.as_slice()) {
+    pub fn import_binary(&mut self, data: PackedByteArray, physics_engine: &mut PhysicsEngine) {
+        match bincode::deserialize::<ShapeImport>(data.as_slice()) {
             Ok(import) => {
-                self.state = import;
+                self.state = import.state;
+                physics_engine.import_shape(import.shape, self.get_handle());
             }
             Err(e) => {
                 godot_error!("Failed to deserialize shape from binary: {}", e);
@@ -170,12 +210,7 @@ impl Drop for RapierShapeBase {
 }
 #[cfg(test)]
 mod tests {
-    use rapier::prelude::*;
-
     use super::*;
-    fn create_test_rigid_body_handle() -> RigidBodyHandle {
-        RigidBodyHandle::from_raw_parts(1, 0)
-    }
     fn create_test_shape_handle() -> ShapeHandle {
         ShapeHandle::from_raw_parts(1, 0)
     }
@@ -201,15 +236,15 @@ mod tests {
     fn test_add_and_remove_owner() {
         let rid = Rid::new(123);
         let mut shape_base = RapierShapeBase::new(rid);
-        let rb_handle = create_test_rigid_body_handle();
-        shape_base.add_owner(rb_handle);
-        assert_eq!(shape_base.get_owners().get(&rb_handle), Some(&1));
-        shape_base.add_owner(rb_handle);
-        assert_eq!(shape_base.get_owners().get(&rb_handle), Some(&2));
-        shape_base.remove_owner(rb_handle);
-        assert_eq!(shape_base.get_owners().get(&rb_handle), Some(&1));
-        shape_base.remove_owner(rb_handle);
-        assert!(shape_base.get_owners().get(&rb_handle).is_none());
+        let rb_id = next_id();
+        shape_base.add_owner(rb_id);
+        assert_eq!(shape_base.get_owners().get(&rb_id), Some(&1));
+        shape_base.add_owner(rb_id);
+        assert_eq!(shape_base.get_owners().get(&rb_id), Some(&2));
+        shape_base.remove_owner(rb_id);
+        assert_eq!(shape_base.get_owners().get(&rb_id), Some(&1));
+        shape_base.remove_owner(rb_id);
+        assert!(shape_base.get_owners().get(&rb_id).is_none());
     }
     #[test]
     fn test_destroy_shape() {
